+++ js/createDate.js
@@ -1,0 +1,5 @@
+//функция создание даты 
+export const createDate=()=>{
+    const date =` ${new Date().toLocaleDateString()} ${new Date().toLocaleTimeString()}`
+    return date
+}
