+++ js/getRandom.js
@@ -1,0 +1,4 @@
+//Функция генератор случайных чисел
+export const getRandom= (num) => {
+    return Math.floor(Math.random() * (num - 1) + 1);
+}
