+++ js/playerLose.js
@@ -1,0 +1,6 @@
+export const  playerLose=(name)=> {
+    const $loseTitle = document.createElement('div');
+    $loseTitle.classList.add('loseTitle');
+    $loseTitle.innerText = `${name} lose`;
+    return $loseTitle;
+}
