+++ js/veriables.js
@@ -1,0 +1,5 @@
+export const $blockMain = document.querySelector('.root');
+export const $blockArena = document.querySelector('.arenas');
+export const $blockRoot = document.createElement('div');
+export const $formFight = document.querySelector('.control');
+export const $chat = document.querySelector('.chat');
