+++ style.css
@@ -205,15 +205,8 @@ body {
     height: 100%;
     position: absolute;
     bottom: 0;
-<<<<<<< Updated upstream
-
     left: 50%;
     transform: translate(-50%, 0);
-
-=======
-    left: 50%;
-    transform: translate(-50%, 0);
->>>>>>> Stashed changes
 }
 
 .fight {
